+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TickCounter.java
@@ -13,20 +13,20 @@ public class TickCounter extends LinearOpMode {
 
     public void runOpMode(){
         Robot robot = new Robot(this.hardwareMap);
-        robot.setDriveMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        robot.stop();
+        robot.drive.setDriveMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        robot.drive.stop();
 
         waitForStart();
-        robot.setDriveMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
+        robot.drive.setDriveMode(DcMotor.RunMode.RUN_WITHOUT_ENCODER);
 
         while(opModeIsActive()){
             robot.activeOpmode = true;
-            telemetry.addData("left Position", (robot.leftFront.getCurrentPosition()+robot.leftBack.getCurrentPosition())/2.0);
-            telemetry.addData("right Position", (robot.rightFront.getCurrentPosition()+robot.rightBack.getCurrentPosition())/2.0);;
+            telemetry.addData("left Position", (robot.drive.leftFront.getCurrentPosition()+robot.drive.leftBack.getCurrentPosition())/2.0);
+            telemetry.addData("right Position", (robot.drive.rightFront.getCurrentPosition()+robot.drive.rightBack.getCurrentPosition())/2.0);;
             telemetry.update();
         }
         robot.activeOpmode = false;
-        robot.stop();
+        robot.drive.stop();
     }
 }
